Move script to bin folder

+++ bin/resize.sh
@@ -5,6 +5,8 @@ start=`date -u +%s`
 start_timestamp=`date -u +'%F %T'`
 scripts=`realpath $0`
 folder=`dirname $scripts`
+autoscaling_folder=$folder/../autoscaling
+monitoring_folder=$folder/../monitoring
 
 if [ $# -eq 0 ]
 then
@@ -43,33 +45,33 @@ if [ $resize_type != "default" ]
 then
   if [ $permanent -eq 0 ]
   then
-    cd $folder/../autoscaling/clusters/$cluster_name
+    cd $autoscaling_folder/clusters/$cluster_name
     cluster_id=`cat cluster_id`
-    shape=`cat inventory | grep shape= | awk -F  "=" '/1/ {print $2}'`
-    queue=`cat inventory | grep queue= | awk -F  "=" '/1/ {print $2}'`
-    log=$folder/../autoscaling/logs/resize_${cluster_id}.log
+    shape=`cat inventory | grep shape= | awk -F  "=" '{print $2}'`
+    queue=`cat inventory | grep queue= | awk -F  "=" '{print $2}'`
+    log=$autoscaling_folder/logs/resize_${cluster_id}.log
     echo $date >> ${log} 2>&1
     if [ -f "currently_resizing" ] && [[ $2 != FORCE ]]
     then
       echo "The cluster is already being resized"
     else
-      echo $1 >> currently_resizing
+      echo $cluster_name >> currently_resizing
       echo `date -u '+%Y%m%d%H%M'` >> $log 2>&1
     fi
   else
     cluster_id=$cluster_name
-    shape=`cat /etc/ansible/hosts | grep shape= | awk -F  "=" '/1/ {print $2}'`
-    queue=`cat /etc/ansible/hosts | grep queue= | awk -F  "=" '/1/ {print $2}'`
-    log=$folder/../autoscaling/logs/resize_${cluster_id}.log
+    shape=`cat /etc/ansible/hosts | grep shape= | awk -F  "=" '{print $2}'`
+    queue=`cat /etc/ansible/hosts | grep queue= | awk -F  "=" '{print $2}'`
+    log=$autoscaling_folder/logs/resize_${cluster_id}.log
   fi
 
-  if [ -f $folder/../monitoring/activated ]
+  if [ -f $monitoring_folder/activated ]
   then
-    source $folder/../monitoring/env
+    source $monitoring_folder/env
     mysqlsh $ENV_MYSQL_USER@$ENV_MYSQL_HOST -p$ENV_MYSQL_PASS --sql -e "use $ENV_MYSQL_DATABASE_NAME; UPDATE cluster_log.clusters SET started_resize='$start_timestamp',state='resizing' WHERE id='$cluster_id'" >> $log 2>&1
   fi
 
-  python3 $folder/../bin/resize.py ${@} >> $log 2>&1
+  python3 $folder/resize.py ${@} >> $log 2>&1
   status=$?
   end=`date -u +%s`
   end_timestamp=`date -u +'%F %T'`
@@ -77,16 +79,16 @@ then
 
   if [ $status -eq 0 ]
   then
-    echo "Successfully Resized cluster $1 in $runtime seconds"
-    if [ -f $folder/../monitoring/activated ]
+    echo "Successfully Resized cluster $cluster_name in $runtime seconds"
+    if [ -f $monitoring_folder/activated ]
     then
-      nodes_list=`python3 $folder/../bin/resize.py --cluster_name $cluster_name list | grep ocid1.instance`
+      nodes_list=`python3 $folder/resize.py --cluster_name $cluster_name list | grep ocid1.instance`
 
       length=`echo $nodes_list | wc -w`
       newSize=$((length/3))
-      existing_nodes=`mysqlsh $ENV_MYSQL_USER@$ENV_MYSQL_HOST -p$ENV_MYSQL_PASS --sql -e "use $ENV_MYSQL_DATABASE_NAME; select hostname from nodes WHERE cluster_id='$cluster_id' and state <> 'deleted ';" 2>&1 | grep inst`
+      existing_nodes=`mysqlsh $ENV_MYSQL_USER@$ENV_MYSQL_HOST -p$ENV_MYSQL_PASS --sql -e "use $ENV_MYSQL_DATABASE_NAME; select hostname from nodes WHERE cluster_id='$cluster_id' and state <> 'deleted';" 2>&1 | grep inst`
       max_index=`mysqlsh $ENV_MYSQL_USER@$ENV_MYSQL_HOST -p$ENV_MYSQL_PASS --sql -e "use $ENV_MYSQL_DATABASE_NAME; select max(cluster_index) from nodes WHERE cluster_id='$cluster_id';" 2>&1 | tail -n 1`
-      mysqlsh $ENV_MYSQL_USER@$ENV_MYSQL_HOST -p$ENV_MYSQL_PASS --sql -e "use $ENV_MYSQL_DATABASE_NAME; UPDATE cluster_log.clusters SET nodes=$newSize,state='running',resize_log='$folder/logs/resize_${cluster_id}.log' WHERE id='$cluster_id'" >> $log 2>&1
+      mysqlsh $ENV_MYSQL_USER@$ENV_MYSQL_HOST -p$ENV_MYSQL_PASS --sql -e "use $ENV_MYSQL_DATABASE_NAME; UPDATE cluster_log.clusters SET nodes=$newSize,state='running',resize_log='$autoscaling_folder/logs/resize_${cluster_id}.log' WHERE id='$cluster_id'" >> $log 2>&1
       if [ $resize_type == "remove" ]
       then
         if [ "$nodes" == "NULL" ]
@@ -123,14 +125,17 @@ then
       fi
     fi
   else
-    echo "Could not resize cluster $1 in 5 tries (Time: $runtime seconds)"
-    if [ -f $folder/../monitoring/activated ]
+    echo "Could not resize cluster $cluster_name in 5 tries (Time: $runtime seconds)"
+    if [ -f $monitoring_folder/activated ]
     then
-      mysqlsh $ENV_MYSQL_USER@$ENV_MYSQL_HOST -p$ENV_MYSQL_PASS --sql -e "use $ENV_MYSQL_DATABASE_NAME; INSERT INTO cluster_log.errors_timeserie (cluster_id,state,error_log,error_type,created_on_m) VALUES ('$cluster_id','resize','$folder/logs/resize_${cluster_id}.log','`tail $log | grep Error`','$end_timestamp');" >> $log 2>&1
+      mysqlsh $ENV_MYSQL_USER@$ENV_MYSQL_HOST -p$ENV_MYSQL_PASS --sql -e "use $ENV_MYSQL_DATABASE_NAME; INSERT INTO cluster_log.errors_timeserie (cluster_id,state,error_log,error_type,created_on_m) VALUES ('$cluster_id','resize','$autoscaling_folder/logs/resize_${cluster_id}.log','`tail $log | grep Error`','$end_timestamp');" >> $log 2>&1
       mysqlsh $ENV_MYSQL_USER@$ENV_MYSQL_HOST -p$ENV_MYSQL_PASS --sql -e "use $ENV_MYSQL_DATABASE_NAME; UPDATE cluster_log.clusters SET started_resizing=NULL,state='running' WHERE id='$cluster_id'" >> $log 2>&1
     fi
+  fi
+  if [ $permanent -eq 0 ]
+  then
     rm currently_resizing
   fi
 else
-  python3 $folder/../bin/resize.py ${@}
+  python3 $folder/resize.py ${@}
 fi
